refactor: Centralize all configuration schemas in config/schemas.py
- Move all trainer configuration classes from trainer/config.py to config/schemas.py
  - Enums: SamplingStrategy, EpochLengthPolicy, ValidationFrequency, ValAggregation
  - Dataclasses: CheckpointConfig, PreemptionConfig, PerformanceConfig, etc.
  - Validation functions: validate_trainer_config, validate_infinite_loader_constraints

- Convert trainer/config.py into a thin re-export wrapper for backward compatibility

- Update package initializers to expose configurations from config.schemas
  - model_training_framework/config/__init__.py now exports trainer configs
  - Maintain backward compatibility through re-exports

- Update all imports across the repository to use config.schemas
  - Updated demo scripts, documentation, and test files
  - Replaced model_training_framework.trainer.config imports with config.schemas

- Add trainer configurations to ExperimentConfig for unified schema

This creates a single source of truth for all configuration objects while
maintaining full backward compatibility for existing code.

All tests pass successfully (298 tests, 283 passed, 15 skipped).
