+++ demo/example1_beginner_local/multi_loader_training.py
@@ -19,6 +19,12 @@ from torch import nn
 import torch.nn.functional as F
 from torch.utils.data import DataLoader, TensorDataset
 
+from model_training_framework.config.schemas import (
+    EpochLengthPolicy,
+    SamplingStrategy,
+    ValAggregation,
+    ValidationFrequency,
+)
 from model_training_framework.trainer import (
     CheckpointConfig,
     GenericTrainer,
@@ -27,12 +33,6 @@ from model_training_framework.trainer import (
     MultiDataLoaderConfig,
     PerformanceConfig,
     ValidationConfig,
-)
-from model_training_framework.trainer.config import (
-    EpochLengthPolicy,
-    SamplingStrategy,
-    ValAggregation,
-    ValidationFrequency,
 )
