+++ model_training_framework/config/__init__.py
@@ -20,45 +20,81 @@ from .naming import (
     ExperimentNaming,
 )
 from .schemas import (
+    # Trainer configuration classes
+    CheckpointConfig,
+    # Core configuration classes
     DataConfig,
+    DDPConfig,
+    EpochLengthPolicy,
+    # Enums
     ExecutionMode,
     ExperimentConfig,
+    FaultToleranceConfig,
+    GenericTrainerConfig,
     GridSearchConfig,
+    HooksConfig,
     LoggingConfig,
     ModelConfig,
+    MultiDataLoaderConfig,
     NamingStrategy,
     OptimizerConfig,
+    PerformanceConfig,
+    PreemptionConfig,
+    SamplingStrategy,
     SchedulerConfig,
     SLURMConfig,
     TrainingConfig,
+    ValAggregation,
+    ValidationConfig,
+    ValidationFrequency,
+    # Validation functions
+    validate_infinite_loader_constraints,
+    validate_trainer_config,
 )
 from .validators import (
     ConfigValidator,
     ValidationResult,
 )
 
-__all__ = [
+__all__ = [  # noqa: RUF022
     # Validation
     "ConfigValidator",
+    "ValidationResult",
+    "validate_infinite_loader_constraints",
+    "validate_trainer_config",
     # Management
     "ConfigurationManager",
-    "DataConfig",
+    # Enums
     "ExecutionMode",
-    # Schemas
+    "NamingStrategy",
+    "SamplingStrategy",
+    "EpochLengthPolicy",
+    "ValidationFrequency",
+    "ValAggregation",
+    # Core configuration schemas
+    "DataConfig",
     "ExperimentConfig",
-    # Naming
-    "ExperimentNaming",
     "GridSearchConfig",
-    "GridSearchResult",
     "LoggingConfig",
     "ModelConfig",
-    "NamingStrategy",
     "OptimizerConfig",
+    "SchedulerConfig",
+    "SLURMConfig",
+    "TrainingConfig",
+    # Trainer configuration schemas
+    "CheckpointConfig",
+    "PreemptionConfig",
+    "PerformanceConfig",
+    "MultiDataLoaderConfig",
+    "ValidationConfig",
+    "FaultToleranceConfig",
+    "DDPConfig",
+    "HooksConfig",
+    "GenericTrainerConfig",
+    # Naming
+    "ExperimentNaming",
     # Grid search
+    "GridSearchResult",
     "ParameterGrid",
     "ParameterGridSearch",
-    "SLURMConfig",
-    "SchedulerConfig",
-    "TrainingConfig",
-    "ValidationResult",
 ]
